+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 **Added:**
+* Flowstep ability, allows airstepping to mobs at the cost of some Focus points (relative to the distance).
+  * Unlocked at level 33 for Warrior, 37 for Mystic and 40 for Guardian.
 * Sound when toggling the selected Reaction Command.
 * Interpolated values for MP Bar while on recharge.
 * Magic level field to Magic Cast Event.